// client/src/Enums/CurrencyEnum.ts
export type CurrencyEnum =
  // Arab Countries
  | "AED" // UAE Dirham
  | "SAR" // Saudi Riyal
  | "QAR" // Qatari Riyal
  | "KWD" // Kuwaiti Dinar
  | "BHD" // Bahraini Dinar
  | "OMR" // Omani Rial
  | "EGP" // Egyptian Pound
  | "LYD" // Libyan Dinar
  | "DZD" // Algerian Dinar
  | "TND" // Tunisian Dinar
  | "MAD" // Moroccan Dirham
  | "SDG" // Sudanese Pound
  | "SOS" // Somali Shilling
  | "DJF" // Djiboutian Franc
  | "IQD" // Iraqi Dinar
  | "SYP" // Syrian Pound
  | "LBP" // Lebanese Pound
  | "YER" // Yemeni Rial
  | "JOD" // Jordanian Dinar

  // Primary Global Currencies
  | "USD" // US Dollar
  | "EUR" // Euro
  | "GBP" // British Pound
  | "CAD" // Canadian Dollar
  | "AUD" // Australian Dollar
  | "NZD" // New Zealand Dollar
  | "CHF" // Swiss Franc
  | "JPY" // Japanese Yen
  | "CNY" // Chinese Yuan
  | "INR" // Indian Rupee
  | "TRY" // Turkish Lira
  | "RUB" // Russian Ruble
  | "BRL" // Brazilian Real
  | "MXN" // Mexican Peso
  | "ZAR"; // South African Rand